ruby marker
