+++ tests/test_the_test/test_group_rules.py
@@ -49,6 +49,7 @@ def test_tracer_release():
         scenarios.k8s_lib_injection_spark_djm,
         scenarios.k8s_lib_injection_uds,
         scenarios.k8s_lib_injection,
+        scenarios.k8s_lib_injection_all_namespaces_old,
         scenarios.k8s_lib_injection_single_service,
         scenarios.lib_injection_validation_unsupported_lang,
         scenarios.lib_injection_validation,
